+++ hoist-error.nix
@@ -1,7 +1,7 @@
 { mkDerivation, base, either, mtl, stdenv }:
 mkDerivation {
   pname = "hoist-error";
-  version = "0.2.0.0";
+  version = "0.2.1.0";
   src = ./.;
   libraryHaskellDepends = [ base either mtl ];
   description = "Some convenience facilities for hoisting errors into a monad";